pyproject.toml
[project]
name = "reflexion-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-ollama (>=0.3.6,<0.4.0)",
    "langgraph (>=0.6.5,<0.7.0)",
    "langchain-tavily (>=0.2.11,<0.3.0)",
    "grandalf (>=0.8,<0.9)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
